correctness of the time

+++ main.py
@@ -19,13 +19,17 @@ logging.basicConfig(
 
 # конфигурационный файл для считывания времени времени выполнения программы
 def file_time():
-    file_path = "/config.txt"
+    file_path = "config.txt"
     # файл найден - получить значение
     if os.access(file_path, os.R_OK):
         config_file = open(file_path)
         run_time = config_file.read()
         config_file.close()
-        return run_time
+        # проверка на корректность значения времени в файле config.txt
+        if run_time.isdigit() == False:
+            return False
+        else:
+            return run_time
     # файл не найден - создать файл, записать значение по умолчанию (5)
     else:
         config_file = open('config.txt', 'w')
@@ -68,22 +72,26 @@ def ping():
         print(text)
 
 if __name__ == "__main__":
-    # создание процессов для параллельной работы
-    timer_process = Process(target=timer)
-    ping_process = Process(target=ping)
-    ip_process = Process(target=get_ip)
-    # запуск процессов
-    timer_process.start()
-    ping_process.start()
-    ip_process.start()
-    # ожидание окончания времени
-    timer_process.join()
-    # проверка на активность процессов ping и получения ip; если процесс не завершен, то завершить его принудительно
-    if ping_process.is_alive():
-        logging.info("Ping process is forcibly terminated")
-        print("Ping process is forcibly terminated")
-        ping_process.terminate()
-    if ip_process.is_alive():
-        logging.info("GetIP process is forcibly terminated")
-        print("GetIP process is forcibly terminated")
-        ip_process.terminate()
+    # проверка на корректность значения времени в файле config.txt
+    if file_time() == False:
+        print("Incorrect value")
+    else:
+        # создание процессов для параллельной работы
+        timer_process = Process(target=timer)
+        ping_process = Process(target=ping)
+        ip_process = Process(target=get_ip)
+        # запуск процессов
+        timer_process.start()
+        ping_process.start()
+        ip_process.start()
+        # ожидание окончания времени
+        timer_process.join()
+        # проверка на активность процессов ping и получения ip; если процесс не завершен, то завершить его принудительно
+        if ping_process.is_alive():
+            logging.info("Ping process is forcibly terminated")
+            print("Ping process is forcibly terminated")
+            ping_process.terminate()
+        if ip_process.is_alive():
+            logging.info("GetIP process is forcibly terminated")
+            print("GetIP process is forcibly terminated")
+            ip_process.terminate()
